Move listener out of cluster
